A major rewrite to the config system. We now use a dictionary of dictionaries internally and store the settings using JSON.
For the end user, this now means including Newtonsoft.json.dll and Newtonsoft.json.xml in the adiIrc package. These can be found in the bin\debug folder of the project.

The rewrite also means we can now bring back the ability to turn on and off server messages for each server. Just toggle speech from the server window for the one you wish to change.